@@ -6,7 +6,7 @@
         "statamic"
     ],
     "homepage": "https://rapidez.io",
-    "license": "GPL-3.0",
+    "license": "GPL-3.0-or-later",
     "authors": [
         {
             "name": "Bob Wezelman",
@@ -30,11 +30,14 @@
     "minimum-stability": "dev",
     "prefer-stable": true,
     "require": {
-        "php": "^8.0|^8.1",
-        "doublethreedigital/runway": "^5.0",
-        "rapidez/blade-directives": "^0.4",
-        "spatie/statamic-responsive-images": "^4.0",
-        "statamic/cms": "^4.0",
+        "php": "^8.1|^8.2",
+        "rapidez/blade-directives": "^0.6",
+        "rapidez/core": "^2.13",
+        "justbetter/statamic-glide-directive": "^2.1",
+        "spatie/once": "*",
+        "statamic-rad-pack/runway": "^7.6",
+        "statamic/cms": "^5.29",
+        "statamic/eloquent-driver": "^4.9",
         "tormjens/eventy": "^0.8"
     },
     "autoload": {
@@ -52,7 +55,10 @@
         "laravel": {
             "providers": [
                 "Rapidez\\Statamic\\RapidezStatamicServiceProvider"
-            ]
+            ],
+            "aliases": {
+                "RapidezStatamic": "Rapidez\\Statamic\\Facades\\RapidezStatamic"
+            }
         }
     }
 }
