@@ -2,7 +2,7 @@
 
 namespace Rapidez\Statamic\Models;
 
-use DoubleThreeDigital\Runway\Traits\HasRunwayResource;
+use StatamicRadPack\Runway\Traits\HasRunwayResource;
 use Illuminate\Database\Eloquent\Model;
 use Statamic\Facades\Site;
 use Statamic\Statamic;
@@ -15,6 +15,9 @@ class Category extends Model
 
     public function getTable()
     {
-        return 'catalog_category_flat_store_'.(Statamic::isCpRoute() ? Site::selected()->attributes['magento_store_id'] : Site::current()->attributes['magento_store_id']);
+        return 'catalog_category_flat_store_'.once(fn() => (Statamic::isCpRoute() 
+            ? (Site::selected()->attributes['magento_store_id'] ?? '1') 
+            : (Site::current()->attributes['magento_store_id'] ?? '1')
+        ));
     }
 }
